requirements.txt
pytest-bdd
allure-pytest-bdd
allure-pytest
playwright
pytest